Drop unnecessary intermediary class
